+++ examples/oscillator/src/lib.rs
@@ -2,15 +2,15 @@ use std::{collections::HashMap, f32::consts::TAU};
 use pure_audio::{MonoSynthData, State};
 
 struct Voice {
-    phase: f32,
-    phase_increment: f32,
+    phase: f64,
+    phase_increment: f64,
     velocity_gain: f32
 }
 
 impl Voice {
     #[inline]
-    fn new(key: u8, velocity: u8 /* 0-127 */, sample_rate: f32) -> Self {
-        let frequency = 440.0 * 2f32.powf((key as f32 - 57.0) / 12.0);
+    fn new(key: u8, velocity: u8 /* 0-127 */, sample_rate: f64) -> Self {
+        let frequency = 440.0 * 2f64.powf((key as f64 - 57.0) / 12.0);
         let phase_increment = frequency / sample_rate;
         let velocity_gain = velocity as f32 / 127.0;
         Self {
@@ -21,7 +21,7 @@ impl Voice {
     }
 
     #[inline]
-    fn advance(&mut self) -> f32 {
+    fn advance(&mut self) -> f64 {
         self.phase += self.phase_increment;
         // avoid overflow on phase
         // sin(1*2pi) = sin(0*2pi) = 0
@@ -34,12 +34,12 @@ impl Voice {
 
 #[derive(Default)]
 pub struct OscillatorState {
-    sample_rate: f32,
+    sample_rate: f64,
     voices: HashMap<u8, Voice>
 }
 
 impl State for OscillatorState {
-    fn activate(&mut self, sample_rate: f32, _min_frame_count: usize, _max_frame_count: usize) {
+    fn activate(&mut self, sample_rate: f64, _min_frame_count: usize, _max_frame_count: usize) {
         self.sample_rate = sample_rate;
     }
 }
@@ -71,7 +71,7 @@ pub fn process(
             voices
                 .values_mut()
                 .fold(0.0f32, |current, voice| 
-                    current + (TAU * voice.advance()).sin() * voice.velocity_gain * gain_per_voice);
+                    current + (TAU * voice.advance() as f32).sin() * voice.velocity_gain * gain_per_voice);
         *sample = sum;
     }
 }
